refactor: extract category tree component
